+++ apps/front-end/public/favicon.ico
