+++ apps/front-end/public/manifest.json
