+++ apps/front-end/src/components/Button.module.css
@@ -1,0 +1,4 @@
+
+.button {
+  background-color: green;
+}
