+++ apps/front-end/src/components/Button.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import '/Button.module.css';
+
+type Props = {
+  title: string;
+};
+
+export const ButtonPrimary = ({ title }: Props) => {
+  return (
+    <button className='button'>{title}</button>
+  );
+}
