+++ apps/front-end/src/screens/Auth/LogIn.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { ButtonPrimary } from '../../components';
+
+// named export
+export const LogIn = () => {
+  return (
+    <div>
+      <h1>Log in page</h1>
+      <ButtonPrimary title='Simple button' />
+    </div>
+  );
+};
