+++ apps/front-end/tsconfig.json
@@ -20,6 +20,6 @@
     "jsx": "preserve"
   },
   "include": [
-    "src"
+    "src/*"
   ]
 }
